requirements.txt
# Python 3.12.0

attrs==23.1.0
beautifulsoup4==4.12.2
black[jupyter]==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
configparser==6.0.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
isort==5.12.0
jsonlines==4.0.0
pandas==2.1.1
pip==23.3.1
praw==7.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
tqdm==4.66.1
urllib3==2.0.7
wheel==0.41.2
wikiextractor==3.0.6
